Refactor lectura de distribucion como text file a otro módulo
